Use common base class for decoder tests
